+++ app/views/suggestions/new.html.erb
@@ -1,6 +1,6 @@
 <div class="box" id="story_box">
   <div id="story_holder">
-    <%= form_with model: @story, url: story_suggest_path(@story.short_id), method: :post, html: { id: 'edit_story' } do |f| %>
+    <%= form_with model: @story, url: suggestions_path(@story.short_id), method: :post, html: { id: 'edit_story' } do |f| %>
       <%= render :partial => "stories/form", :locals => { :story => @story,
         :f => f, :suggesting => true } %>
